pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "keepassxc-ssh-agent"
version = "0.1.0"
description = "SSH IdentityAgent proxy that triggers KeePassXC database unlock via TouchID"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-2.0-or-later"
dependencies = [
    "PyNaCl>=1.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
keepassxc-ssh-agent = "keepassxc_ssh_agent.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
